Wizard form in issues#new using Turbo Streams and Stimulus
Because we want to have each step in a different part of the page, using
Turbo Frame wasn't as straightforward as I initially thought (but we'll
see later, I may have a plan…)

Instead, each step of the form makes its own request to a different
controller, so as to update the main form. The requests are dispatched
by as Stimulus controller, and the response is processed by Turbo Streams.