Bringing sanity back to the palette
Mixing the two attempts at a sane palette (one in `base.css`, the other
in `variables.css`) only led to more confusion. Hopefully, things are
better now (though not yet application-wide).

The main issue was trying to combine two hue gradients – an off-white to
gray and a cream to caramel) into a single namespace, with the `secondary`
one as an extra. Now there are 4 hues in the palette:

- Neutral goes from off-white to dark grey.
- Primary goes from cream to caramel to chocolate (hummy!)
- Secondary is the teals and blues (aka briefs colors)
- Tertiary will be the pinks (aka bloomers colors)