Finally happy with the design & code for slots#new
It works on Safari and Chrome, and in acceptable in Firefox (once CSS
nesting is enabled). Compared with the last commit, the fields layout
on the 3rd step is alright, and the grid on the 2nd step is a bit
different (because issues tend to require more horizontal space than
repo names).